requirements.txt
PyPDF2>=3,<4
python-docx>=1,<2
openai>=1,<2
python-dotenv>=1,<2
faiss-cpu>=1,<2
numpy>=1,<3
Flask>=3,<4
gunicorn>=22,<24
